separated demos from automated tests
